refactor : reducer function and pages
