+++ src/context/cartContext.js
@@ -1,6 +1,6 @@
 import React ,{createContext,useContext, useReducer,useState,useEffect} from 'react'
 import axios from 'axios';
-
+import {CartReducer} from '../Reducer/reducerAll'
 const CartContext= createContext();
 
 const CartProvider=({children})=>{
@@ -18,29 +18,6 @@ const CartProvider=({children})=>{
     }
     fetchData();} , []);
 
-    const CartReducer=(state,action)=>{
-        
-        switch(action.type){
-
-            case"ADD_TO_CART": 
-            return{...state,cart:[...state.cart,{...action.payload,qty:1}]}
-            case"DELETE_FROM_CART":
-            return{...state,cart:state.cart
-                .filter((c)=>c.id!==action.payload.id)};
-            case"ADD_QUANTITY":
-            return {...state,cart:state.cart
-                .filter((c)=>c.id===action.payload.id?c.qty=action.payload.qty+1:c.qty)};
-            case"REDUCE_QUANTITY":
-            return {...state,cart:state.cart
-                .filter((c)=>c.id===action.payload.id?c.qty=action.payload.qty-1:c.qty)
-                .filter(((c)=>c.qty!==0))};
-            case"CHANGE_QUANTITY":
-            return {...state,cart:state.cart
-                .filter((c)=>c.id===action.payload.id?c.qty=action.payload.qty:c.qty)};
-            default:return state;
-        }
-    }
-
     const [state,dispatch]=useReducer(CartReducer,{products: data,cart:[]})
 
     return (
@@ -48,8 +25,6 @@ const CartProvider=({children})=>{
             {children}
         </CartContext.Provider>
     )
-
-
 }
 const useCart = () => useContext(CartContext);
