+++ src/pages/Wishlist/Wishlist.jsx
@@ -5,10 +5,10 @@ import "./Wishlist.css"
 export default function Wishlist() {
   return (
     <div>
-        <div class="ec-first-container">
+        <div className="ec-first-container">
         <NavigationTop/>
-        <div class="ec-wishlist-container">
-            <div class="h3 center-text ec-head-text">My wishlist!</div>
+        <div className="ec-wishlist-container">
+            <div className="h3 center-text ec-head-text">My wishlist!</div>
         <WishlistCard/>
         </div>
     </div>
