+++ src/pages/pagesAll.js
@@ -1,0 +1,7 @@
+import LandingPage from "./Home/LandingPage";
+import Product from "./Product/Product";
+import Wishlist from "./Wishlist/Wishlist";
+import Cart from "./Cart/Cart";
+import PageNotFound from "./PageNotFound";
+
+export{LandingPage,Product,Wishlist,Cart,PageNotFound}
